requirements.txt
streamlit
torch
torchvision
pillow
matplotlib
numpy
scikit-learn
